Migrates SplitOrJoinedTripsValidator to TripTimesValidator interface
